+++ addons/docs/src/frameworks/react/config.js
@@ -7,6 +7,7 @@ addParameters({
     container: DocsContainer,
     page: DocsPage,
     // react is Storybook's "native" framework, so it's stories are inherently prepared to be rendered inline
-    prepareForInline: (storyFn: StoryFn) => storyFn(),
+    // NOTE: that the result is a react element. Hooks support is provided by the outer code.
+    prepareForInline: storyFn => storyFn(),
   },
 });