+++ examples/official-storybook/stories/addon-docs/csf-with-mdx-docs.stories.js
@@ -4,7 +4,7 @@ import { Button } from '@storybook/react/demo';
 export default {
   title: 'Addons|Docs/csf-with-mdx-docs',
   component: Button,
-  includeStories: [], // or simply don't load this file at all
+  includeStories: [], // or don't load this file at all
 };
 
 // eslint-disable-next-line react/prop-types